更新樣式

+++ comments/views.py
@@ -1,4 +1,5 @@
-from django.shortcuts import redirect, get_object_or_404
+from django.shortcuts import redirect, get_object_or_404, render
+from django.http import HttpResponse
 from .forms import CommentForm
 from articles.models import Article
 from django.contrib import messages
@@ -14,15 +15,18 @@ def create(request, id):
     comment = form.save(commit=False)
     comment.article = article
     comment.save()
+    # messages.success(request, "新增留言成功")
+    return render(request, "comments/comment.html", {
+        "comment": comment
+    })
 
-    messages.success(request, "新增留言成功")
-    return redirect("articles:detail", article.id)
-
-
-@require_POST
 def delete(request, id):
-    comment = get_object_or_404(Comment, pk=id)
-    comment.delete()
+    if request.method == "DELETE":
+        comment = get_object_or_404(Comment, pk=id)
+        comment.delete()
 
-    messages.warning(request, "刪除成功")
-    return redirect("articles:detail", comment.article_id)
+        return HttpResponse("")
+
+        # messages.warning(request, "刪除成功")
+
+        # return redirect("articles:detail", comment.article_id)
